Reworking character generation
Reworked so that character is not generated automatically. First you have to decide what attributes to generate, and then click the "Generate Character" button. The front end for this will change.

Re-rolling therefore only happens when you click one of the attributes on the character sheet.

Also reworked index.html so that it implements more Javascript & AJAX to handle character generation.