Removing GPT Name as separate attribute
When using GPT to generate a name, the resulting name is split into first name and last name, instead of being stored as a separate gpt_name.
Also reworked the Dall-E prompt in order to reflect these changes, and fixed information not coming through to the prompt from the front end.