Rotating gacha capsule
Now to generate a character portrait you can click on an image of a gachapon capsule, which then starts rotating. When the image has been generated it replaces the capsule. If you click the image then it hides the portrait, shows the spinning capsule again, until a new portrait has been generated.